requirements.txt
# --- Core Backend ---
fastapi==0.110.0
uvicorn[standard]==0.27.1

# --- WebSocket & Async
websocket==12.0

# --- Audio Processing ---
numpy==1.26.4
pyttsx3==2.90

# --- Vision ---
opencv-python

# --- Utils ---
python-dotenv==1.0.1
pydantic==2.6.4
PyYAML==6.0.1